Set up the grid container and position elements
We set up explicit columns and rows and by using grid-area we have
positioned the three main elements into that grid container.
For now I have set a width and height on the body element of 100vh and
100vw to force the body to use the entire viewport and to not base
the height and width on the intrinsic value of the content in those
elements as they are not full yet.