Add styling to sidebar elements
I changed the font color to white, removed the text decoration white
color.
Instead of adding padding or margin between an anchors image/svg and
h2 text content, I will be using grid and using gap, and the gap amount
will be the same between the svg/img and the h2 as the amount between
each li of the ul in these sidebar elements.
I also just realized why my grid styles have not been working, it's
because I have not been applying them to the ul's but to the ul's
parent.