Add grid to sidebar elements and their children
I have applied a grid to the sidebar elements children and set their
columns to 1, thus the three child elements will all be in one column,
and they'll be spaced apart by a gap of 25px.
I have also applied grid to the anchor elements in the sidebar as they
all wrap two elements a svg and a header and we want to space them out
from each other but on the same line thus we used two columns.
I also removed some test code I had on the main-sidebar that I was
comparing different values to properties.