Change header grid area, apply grid to header
I made a mistake earlier which was that I had the .main-header
positioned to take up all the rows, not just the first row, oh well
simple little fix.
I also applied a background color to the header, applied grid to its
display layout, and gave it 2 columns and 2 rows to position its child
elements into.