Add grid to user-actions, style buttons, etc
In addition to adding grid to user-actions, we've stated that the grid
should automatically make columns and adjust their width/sizing to
however many child elements there are and gave a minimum and maximum
size that will scale dynamically.
I also styled the buttons in user-action as I thought it needed it now
more than later.
As I was aligning the buttons within their parent grid container, I
realized that all grid elements should be aligned center, or at least
at this time I feel that way, so I applied that as well.