Add/change styling to .projects children
Added padding to the .project-article elements, and a very small gap to
separate its children between each other.
Aligned the .project-title appropriately and changed its font weight;
Changed the .project-text to also match the design of gray.
Changed the position of .actions although now that I look at it I think
it should be justify-content, but I'll touch on it later when I add the
icons/logos/etc.