Change padding, and add child layouts
This is in regards to the announcements section.
I wanted to make sure that only one column was in the .announcements-container, maybe unneeded but useful practice, and did the same thing to announcement-article.
Used a pseudo class selector to not apply a bottom border on the third article but instead to the first two articles, as the design wanted.
Just as before moved the h4 which is kind of like the title of a specific section to the bottom of it's grid cell to match desired design layout, and changed its font weight to communicate its importance as I used a css reset which changed the browsers default weight to headings.
The paragraph text in these announcement articles is smaller than the text in the projects section, so I changed the font-weight accordingly.