Add link element for font into head, add img class
Had to add a link element for a Google font for this project, instead
of adding more files to the project I'll just use the link element to
load them in when the page is loaded.
Also added a class for the project-actions img's, next I'll give these
elements svg files to display in their positions.