Remove gap and add styling to welcome
The grid children were spaced out too much so I removed the gap.
To more align the span and h4 in the welcome-text I applied grid to it
and have attempted to position it with align-content but I'm not sure
that I accomplished it the way I want.
I also changed the font sizes of the span and the h4 in the
.welcome-text.