Change padding on header, change header stylings
I had too much padding, this was because of the changes I made to the
grid children.
To also further place child elements of header I needed to apply width
nd heights of 100% so that I could move those elements within the extra
space in their parents grid area.
I also changed some values given to style the .user-controls as they
were not lining up with the search parent container, now the 2 parents
are centered.