Adjust grid columns in the projects container
This is to make it more dynamic, allowing for a dynamic number of
project cards to be displayed in each row of the projects container.
Also adjusted the h3 grid placement using a -1 in the column end so
that if the container shrinks or grows that h3 will still take up the
entire row , aka all of the columns on its row disallowing cards to be
auto placed beside it.